Test does not need JenkinsRule so can be in core/src/test/java/
